v0.67 Menu and interactivity setup in the boss level. Ready to implement mechanics Training level to be done

+++ Assets/Scripts/Player/Abilities/RushAbility.cs
@@ -106,7 +106,7 @@ public class RushAbility : MonoBehaviour {
             _numCharges = Mathf.Clamp(_numCharges++, 0, _maxCharges);
             _numMoths = (_numCharges == _maxCharges ? 1 : 0);
         }
-        _gameHud.SetCooldown(_numMoths / _numMothsToRecharge);
+        _gameHud.SetCooldown((float)_numMoths / _numMothsToRecharge);
     }
 
     private void SetupHudBar()
